requirements.txt
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
chardet==5.2.0
click==8.2.1
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
et_xmlfile==2.0.0
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
